CHORE: adicionado mais validacoes de cliente

+++ GTI.Application/Services/ClienteService.cs
@@ -27,22 +27,29 @@ namespace GTI.Application.Services
         public async Task<CommandResult> GetAllClientes()
         {
             var listaClientes = await _readRepository.FindAll().Include(x => x.Enderecos).ToListAsync();
+
+            if (listaClientes is null || listaClientes.Count == 0)
+                return new CommandResult(false, "Falha ao consultar clientes");
+
             return new CommandResult(true, "Clientes consultados com sucesso", listaClientes);
         }
 
         public async Task<CommandResult> GetByIdCliente(Guid id)
         {
             var cliente = await _readRepository.FindByCondition(x => x.Id == id).Include(x => x.Enderecos).ToListAsync();
+
+            if (cliente is null || cliente.Count == 0)
+                return new CommandResult(false, "Falha ao consultar cliente");
+
             return new CommandResult(true, "Cliente consultado com sucesso", cliente);
         }
 
         public async Task<CommandResult> CreateCliente(CreateClienteCommand command)
         {
-            // TODO FAST FAIL VALIDATION
             command.Validate();
 
             if (!command.IsValid)
-                return new CommandResult(false, "teste", command.Notifications);
+                return new CommandResult(false, "Falha ao registrar cliente", command.Notifications);
 
             Cliente cliente = new Cliente(command.Cpf, command.Nome, command.Rg, command.DataExpedicao, command.OrgaoExpedicao,
                 command.Uf, command.DataDeNascimento, command.Sexo, command.EstadoCivil);
@@ -59,12 +66,20 @@ namespace GTI.Application.Services
 
         public async Task<CommandResult> UpdateCliente(UpdateClienteCommand command)
         {
-            // TODO FAST FAIL VALIDATION
+            command.Validate();
 
-            Cliente cliente = await _readRepository.FindByCondition(x => x.Id == command.IdClienteExistente).FirstOrDefaultAsync();
+            if (!command.IsValid)
+                return new CommandResult(false, "Falha ao alterar cliente", command.Notifications);
+
+            var cliente = await _readRepository.FindByCondition(x => x.Id == command.IdClienteExistente).FirstOrDefaultAsync();
+            if (cliente is null)
+                return new CommandResult(false, "Falha ao recuperar cliente");
+
+            var endereco = await _readEnderecoRepository.FindByCondition(x => x.Cliente.Id == command.IdClienteExistente).FirstOrDefaultAsync();
+            if (endereco is null)
+                return new CommandResult(false, "Falha ao recuperar endereco");
+
             cliente.Alterar(command);
-
-            Endereco endereco = await _readEnderecoRepository.FindByCondition(x => x.Cliente.Id == command.IdClienteExistente).FirstOrDefaultAsync();
             endereco.Alterar(command.Endereco);
 
             _writeRepository.Update(cliente);
@@ -76,10 +91,13 @@ namespace GTI.Application.Services
 
         public async Task<CommandResult> DeleteCliente(Guid id)
         {
-            // TODO FAST FAIL VALIDATION
+            var cliente = await _readRepository.FindByCondition(x => x.Id == id).FirstOrDefaultAsync();
+            if (cliente is null)
+                return new CommandResult(false, "Falha ao recuperar cliente");
 
-            Cliente cliente = await _readRepository.FindByCondition(x => x.Id == id).FirstOrDefaultAsync();
-            Endereco endereco = await _readEnderecoRepository.FindByCondition(x => x.Cliente.Id == id).FirstOrDefaultAsync();
+            var endereco = await _readEnderecoRepository.FindByCondition(x => x.Cliente.Id == id).FirstOrDefaultAsync();
+            if (endereco is null)
+                return new CommandResult(false, "Falha ao recuperar endereco");
 
             _writeRepository.Delete(cliente);
             _writeEnderecoRepository.Delete(endereco);
